Add option to disable self management
The existing implementation will self-manage the symbol database. However, this
means the prototype we use are generated in addition to the ones generated at
build-time. To be more efficient, and follow a more "single source of truth"
model, we should use the ones generated at build-time.

This commit changes the behavior to default to using the ones generated at
build-time; although to do this, we have to make use of the `_sym_db`
attribute, which is, in theory, private. Regardless, we keep the self-managed
option as a backup.